Added example for test case 3
There is not an example for test case 3 so that once the code is complete it should pass if the code is written correctly. I may not have time to get to that feature of the code, but at least I can show that I did consider that case and chose how I think the code should handle it.