Added negative test case and case sensitive test cases
I have run out of time and need to start working on my writeup, so I added the remaining test cases which were explicitly asked for in the instructions. I believe my code works properly for the negative test case and the case sensitive test case, but does not work if there are multiple matches or if the word is hyphenated across lines.